requirements.txt
# Core AI/ML Libraries
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# API and Web
requests>=2.31.0
python-dotenv>=1.0.0

# Jupyter Support
jupyter>=1.0.0
notebook>=7.0.0

# Utilities
pyyaml>=6.0.0
